requirements.txt
aiodns==3.0.0
aiofiles==0.6.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiomysql==0.0.22
aiosignal==1.2.0
aiosqlite==0.17.0
arrow==1.2.2
astunparse==1.6.3
async-cse==0.3.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
braceexpand==0.1.7
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
coloredlogs==15.0.1
colorgram.py==1.2.0
cycler==0.11.0
durations-nlp==1.0.1
flake8==4.0.1
fonttools==4.29.1
frozenlist==1.3.0
humanfriendly==10.0
humanize==3.14.0
idna==3.3
import-expression==1.1.4
jishaku==2.5.0
kdtree==0.16
kiwisolver==1.3.2
matplotlib==3.5.1
mccabe==0.6.1
multidict==6.0.2
numpy==1.22.2
oauthlib==3.2.0
orjson==3.6.6
packaging==21.3
Pillow==9.0.1
psutil==5.9.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyMySQL==0.9.3
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
random-user-agent==1.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
statcord.py==3.1.0
tweepy==3.10.0
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
uvloop==0.16.0
webencodings==0.5.1
yarl==1.7.2
pre-commit==2.18.1

discord.py[speed] @ git+https://github.com/Rapptz/discord.py.git